Add files via upload
Sapphos database work,
Sapphos was an application for relationships. It had a faulty database that compromised the privacy of many users and committed several illegal violations. In this project, we redesigned the ER models to fix the logical issues and then created a new database using Postgres. We created several triggers and procedures to reinforce security.